Procman passes back errors instead of just logging.
-Procman now returns STAT_RUNTIME_ERROR as well as the error string
instead of just logging runtime errors (only works in foreground mode)
-test.lua now has '-e' option that causes it to purposly throw a runtime
error
	modified:   kernel/base/procman/procman.lua
	modified:   kernel/usr/programs/test.lua

+++ kernel/usr/programs/test.lua
@@ -4,21 +4,29 @@ print()
 print("Something not working?")
 print("Never be afraid to try something new. Remember, amateurs built the ark; professionals built the titanic.")
 
+local mode = ...
 local count = 3
-local function thinkfunc()
-    count = count - 1
-    print(tostring(count))
-    if count <= 0 then
-        print("If its any consolation though, I'm working!")
-        print()
-        print()
-        procman.terminate()
+local procinfo =  {
+    ["name"] = "test.lua"
+}
+
+if mode == "-e" then --throw error
+    procinfo["main"] = function()
+        local errorz = nonExistantVariable["meep"]
     end
+else
+    procinfo["hooks"] = {
+        ["think"] =  function()
+            count = count - 1
+            print(tostring(count))
+            if count <= 0 then
+                print("If its any consolation though, I'm working!")
+                print()
+                print()
+                procman.terminate()
+            end
+        end
+    }
 end
 
-return {
-    ["name"] = "test.lua",
-    ["hooks"] = {
-        ["think"] =  thinkfunc
-    }
-}
+return procinfo
